refactor: move data to separate file
