+++ src/App.tsx
@@ -5,6 +5,7 @@ import AboutPage from './pages/AboutPage/AboutPage';
 import Notfoundpage from './pages/NotfoundPage/Notfoundpage';
 import Header from './components/Header';
 import FormsLayout from './components/FormsLayout/FormsLayout';
+import sneakers from './data';
 
 export interface Sneaker {
   id: string;
@@ -15,65 +16,6 @@ export interface Sneaker {
 }
 
 const App = () => {
-  const sneakers = [
-    {
-      id: '7',
-      parentId: '6',
-      title: 'Sneaker Black Edition',
-      price: 118,
-      imageUrl: 'img/sneakers/6.svg',
-    },
-    {
-      id: '9',
-      parentId: '5',
-      title: 'Sneaker Future Rider',
-      price: 139,
-      imageUrl: 'img/sneakers/5.svg',
-    },
-    {
-      id: '10',
-      parentId: 3,
-      title: 'Sneaker Nike Blazer Mid Suede',
-      price: 164,
-      imageUrl: 'img/sneakers/3.svg',
-    },
-    {
-      id: '11',
-      parentId: 4,
-      title: 'Sneaker Puma X Aka Boku Future Rider',
-      price: 123,
-      imageUrl: 'img/sneakers/4.svg',
-    },
-    {
-      id: '12',
-      parentId: 1,
-      title: 'Sneaker Nike Blazer Mid Suede',
-      price: 100,
-      imageUrl: 'img/sneakers/1.svg',
-    },
-    {
-      id: '13',
-      parentId: 4,
-      title: 'Sneaker Puma X Aka Boku Future Rider',
-      price: 123,
-      imageUrl: 'img/sneakers/4.svg',
-    },
-    {
-      id: '16',
-      parentId: 4,
-      title: 'Sneaker Puma X Aka Boku Future Rider',
-      price: 123,
-      imageUrl: 'img/sneakers/3.svg',
-    },
-    {
-      id: '18',
-      parentId: 4,
-      title: 'Sneaker Puma X Aka Boku Future Rider',
-      price: 123,
-      imageUrl: 'img/sneakers/1.svg',
-    },
-  ];
-
   return (
     <div className="container">
       <Header />
